transplanting the VQE algorithm to a new .py file
